# openapi.yaml
openapi: 3.1.0
info:
  title: Monetiser — Payment Orchestration Platform API
  version: 2.11.0
  summary: API real de Monetiser, generada leyendo el código (no una API idealizada).
  description: |
    Contrato REAL de la API de Monetiser, extraído de los DTOs, validators y
    controllers del repositorio `marcmonf/orquestacion-def-test`. Cada campo de
    esta spec existe en el código; no hay campos inventados ni aspiracionales.

    **v2.11.0 (26 sep 2026)** — Vuelta del comprador tras pagar y sesiones
    revocables (ver DEV-LOG, sesión 26 sep 2026):
      - **Portal: sesiones revocables.** Un token deja de valer en cuanto el
        usuario se desactiva, cambia de rol o de nodo, cambia su contraseña o cierra
        sesión (`401 session_revoked`). `POST /portal/auth/logout` ya no es
        "stateless": invalida los tokens del usuario.
      - **Portal: un admin restringido a un nodo** solo lista y gestiona usuarios de
        su subárbol, no puede cambiar su propio nodo (`409 cannot_change_own_node`)
        ni dejar a nadie sin restricción (`403 outside_your_scope`).
        `POST /portal/users` acepta `hierarchyNodeId` (por defecto, el nodo del
        admin si está restringido).
      - Al terminar el pago (3DS incluido) el comprador llega a la **página de
        resultado** de Monetiser (`GET /checkout/result/{paymentId}`, URL firmada que
        Paylands recibe como `url_ok`/`url_ko`). Antes se quedaba en una página de
        Paylands dentro del iFrame.
...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - ApiKeySimple: []
    MerchantId: []
paths:
...
  /portal/auth/login:
    post:
      tags: [Portal]
      summary: Login del portal del merchant
      operationId: portalLogin
      x-madurez: estable
      security: []
      description: |
        Autentica a un usuario de merchant y devuelve un JWT de portal. **Rate
        limited** por IP + email (`RL_PORTAL_LOGIN_MAX`, por defecto 10/15 min):
        superado el límite responde `429`.

        Si `mustChangePassword` es `true`, el token sirve únicamente para
        `POST /portal/auth/change-password`; el resto del portal responde `403`
        hasta cambiar la password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortalLoginRequest'
      responses:
        '200':
          description: Login correcto.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalLoginResponse'
        '400':
          description: Faltan credenciales.
        '401':
          description: Credenciales inválidas o usuario inactivo (`invalid_credentials`).
        '429':
          description: Demasiados intentos (`rate_limit_exceeded`).
  /portal/auth/change-password:
    post:
      tags: [Portal]
      summary: Cambiar la propia password (limpia mustChangePassword)
      operationId: portalChangePassword
      x-madurez: estable
      security: [{ PortalBearer: [] }]
      description: |
        Cambia la password del usuario de la sesión. Permitido incluso bajo
        `mustChangePassword` (es como se limpia el flag en el primer login).
        Devuelve un token fresco con el flag ya a `false`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortalChangePasswordRequest'
      responses:
        '200':
          description: Password actualizada. Devuelve un token nuevo.
        '400':
          description: Falta algún campo, la nueva es menor de 8 caracteres o igual a la actual.
        '401':
          description: Sin token, token inválido, o `currentPassword` incorrecta.
  /portal/auth/logout:
    post:
      tags: [Portal]
      summary: Cerrar sesión (invalida el token)
      operationId: portalLogout
      x-madurez: estable
      security: [{ PortalBearer: [] }]
      description: |
        Invalida TODOS los tokens del usuario de la sesión (sube su versión de
        sesión). Responde `200` siempre, también sin token o con uno caducado.
        Antes era "stateless": el token seguía valiendo hasta caducar (12 h).
      responses:
        '200':
          description: Sesión cerrada.
...
components:
  securitySchemes:

    ApiKeySimple:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        Modo simple, pensado para Postman y pruebas.

        **Solo funciona si `API_KEY_SIMPLE_FALLBACK=true` en el entorno.** Si no lo
        está, hay que usar `HmacV1`.

        **En este header va el `rawSecret` (`ms_...`), NUNCA el `rawKeyId`
        (`mk_...`).** `validateApiKey()` en `src/services/apiKeyService.js` busca
        por `secretHash` (SHA-256 del secreto), igual que el modo HMAC.

        > **Cambio de contrato — 4 ago 2026.** Hasta esa fecha este header
        > llevaba el `rawKeyId` y el `rawSecret` no intervenía en ningún momento:
        > el identificador PÚBLICO de la credencial (visible en `/admin`, en la
        > cabecera `Authorization` del modo HMAC y en los logs) actuaba como
        > contraseña. Corregido. También se retiró el fallback por `keyHash`
        > (SHA-256 del `keyId`), que era el mismo agujero con un hash encima.
        >
        > Quien mande el `keyId` recibe `401` con
        > `detail: "x_api_key_must_be_the_secret_not_the_key_id"`.
        >
        > Sigue siendo un bearer secret en claro: `HmacV1` continúa siendo el
        > modo recomendado en producción, porque además firma la petición.

        Requiere acompañarlo del header `x-merchant-id`.

    MerchantId:
      type: apiKey
      in: header
      name: x-merchant-id
      description: |
        Merchant sobre el que se opera.

        En las rutas `/{merchantId}/...` se puede omitir: se toma de la URL. En las
        de ciclo de vida (`/payments/{paymentId}/...`) **es obligatorio**, porque el
        merchant no viaja en la ruta.
...
  schemas:
...
    PortalLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:    { type: string, format: email }
        password: { type: string }
    PortalLoginResponse:
      type: object
      properties:
        success:            { type: boolean, const: true }
        token:              { type: string, description: JWT de portal (aud portal). }
        mustChangePassword: { type: boolean }
        user:
          type: object
          properties:
            email:      { type: string, format: email }
            name:       { type: string }
            role:       { type: string }
            merchantId: { type: string }
    PortalChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string }
        newPassword:     { type: string, minLength: 8, description: Distinta de la actual. }
...